Refactor usersService
